pyproject.toml
[project]
name = "mcp-template"
version = "0.1.0"
description = "A template MCP server for quickly starting a new project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dedalus-mcp",
    "pydantic",
    "python-dotenv",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build]
targets.wheel.packages = ["src"]
